[API](SR) Rename performance-tools.js to monitoring-tools.js
